chore(apps/web): separate out indexeddb types from tinybird files.
