+++ packages/tinybird/src/publish.ts
@@ -30,7 +30,6 @@ const tinybirdEmail = z.object({
   sizeEstimate: z.number().nullish(), // Estimated size in bytes
 });
 export type TinybirdEmail = z.infer<typeof tinybirdEmail>;
-export type IndexedDBEmail = z.infer<typeof tinybirdEmail>;
 
 export const publishEmail = tb.buildIngestEndpoint({
   datasource: "email",
